Update utils.py
Change read-in data from int to float

+++ code/utils/utils.py
@@ -13,7 +13,7 @@ def load_dataset(filename):
         num_items, capacity = map(int, file.readline().strip().split())
         
         for _ in range(num_items):
-            value, weight = map(int, file.readline().strip().split())
+            value, weight = map(float, file.readline().strip().split())
             items.append({'value': value, 'weight': weight})
 
     return items, capacity
